# Significant performance increase in bottom up inference by reducing GPU to CPU data transfer. 